requirements.txt
streamlit
Pillow
rembg
pandas
cv2